requirements.txt
flask
flask-cors
requests
pymongo
pyjwt